// front/app/jeddah/model/mdlJeddahParams.ts
export interface MdlJeddahInputxAllpnr {
  chssmr: string;
  clssfl_pnrdtl: string;
  airlfl_pnrdtl: string;
  datefl_pnrdtl: string;
  pnrcde_pnrdtl: string;
  pnrclk_pnrdtl: string;
  flnbfl_pnrdtl: string;
  routfl_pnrdtl: string;
  agtnme_pnrdtl: string;
  srtspl_pnrdtl: string;
  srtcxl_pnrdtl: string;
  pagenw_pnrdtl: number;
  limitp_pnrdtl: number;
  airlfl_pnrsmr: string;
  flnbfl_pnrsmr: string;
  routfl_pnrsmr: string;
  datefl_pnrsmr: string;
  pnrcde_pnrsmr: string;
  pnrclk_pnrsmr: string;
  agtnme_pnrsmr: string;
  psdate_pnrsmr: string;
  srtspl_pnrsmr: string;
  srtcxl_pnrsmr: string;
  pagenw_pnrsmr: number;
  limitp_pnrsmr: number;
  airlfl_flnsmr: string;
  datefl_flnsmr: string;
  flnbfl_flnsmr: string;
  routfl_flnsmr: string;
  psdate_flnsmr: string;
  srtspl_flnsmr: string;
  srtcxl_flnsmr: string;
  pagenw_flnsmr: number;
  limitp_flnsmr: number;
  airlfl_agtnme: string;
  agtnme_agtnme: string;
  srtnul_agtnme: string;
  pagenw_agtnme: number;
  limitp_agtnme: number;
}
// ...
export interface MdlJeddahSearchPnrsmr {
  airlfl_pnrsmr: string;
  flnbfl_pnrsmr: string;
  routfl_pnrsmr: string;
  datefl_pnrsmr: string;
  pnrcde_pnrsmr: string;
  agtnme_pnrsmr: string;
  psdate_pnrsmr: string;
  srtspl_pnrsmr: string;
  srtcxl_pnrsmr: string;
  pagenw_pnrsmr: number;
  limitp_pnrsmr: number;
}